:recycle: [ProfessionalController] Move list/filter route in BeneficiaryController
