+++ templates/v2/beneficiary/list/_beneficiaries_list.html.twig
@@ -2,13 +2,13 @@
 
 {% block listRows %}
     {% for beneficiary in beneficiaries %}
-        {{ include('v2/professional/_beneficiaries_list_row.html.twig', {beneficiary}) }}
+        {{ include('v2/beneficiary/list/_beneficiaries_list_row.html.twig', {beneficiary}) }}
     {% endfor %}
 {% endblock %}
 
 {% block pagination %}
     {{ knp_pagination_render(beneficiaries, 'v2/vault/components/_pagination.html.twig', {}, {
-        'routeName': 'search_beneficiaries',
+        'routeName': 'filter_beneficiaries',
         'id': null
     }) }}
 {% endblock %}
