+++ templates/v2/beneficiary/list/_beneficiaries_list_row.html.twig
@@ -20,7 +20,7 @@
 {% endblock %}
 
 {% block vaultItems %}
-    {{ include('v2/professional/_personal_data_row.html.twig') }}
+    {{ include('v2/beneficiary/list/_personal_data_row.html.twig') }}
 {% endblock %}
 
 {% block actionDropdown %}
