+++ templates/v2/beneficiary/list/_personal_data_row.html.twig
