+++ templates/v2/beneficiary/list/beneficiaries.html.twig
@@ -20,5 +20,5 @@
 {% endblock %}
 
 {% block list %}
-    {{ include('v2/professional/_beneficiaries_list.html.twig', {beneficiaries}) }}
+    {{ include('v2/beneficiary/list/_beneficiaries_list.html.twig', {beneficiaries}) }}
 {% endblock %}
