+++ tests/v2/Controller/Beneficiary/ListBeneficiariesTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Tests\v2\Controller\ProController;
+namespace App\Tests\v2\Controller\Beneficiary;
 
 use App\DataFixtures\v2\BeneficiaryFixture;
 use App\DataFixtures\v2\MemberFixture;
